fix: C2 fp16 RoPE + C5 eval length cap
- rotary cos/sin tables cast to the activation dtype (fp16-safe under AMP;
  fp32 promotion made the RoPE path behave differently on GPU - the only
  code-level difference between C1 and C2)
- BLT eval capped at window length (byte-mode has no <eos>, so greedy
  decoding always emitted 256 bytes against 32-byte window references;
  bit_accuracy penalised the padded bytes -> ceiling ~12.5% despite the
  best val loss of the suite). Cap now 40 bytes (32-char windows + slack)

+++ src/models/positional.py
@@ -39,8 +39,8 @@ class RotaryEmbedding(nn.Module):
         self.register_buffer("cos", torch.cos(angles).repeat_interleave(2, dim=-1).view(1, 1, max_len, d_k))
         self.register_buffer("sin", torch.sin(angles).repeat_interleave(2, dim=-1).view(1, 1, max_len, d_k))
 
-    def rotate(self,q, k):
+    def rotate(self, q, k):
         s = q.size(2)
-        cos = self.cos[:, :, :s]
-        sin = self.sin[:, :, :s]
+        cos = self.cos[:, :, :s].to(q.dtype)
+        sin = self.sin[:, :, :s].to(q.dtype)
         return q * cos + _rotate_half(q) * sin, k * cos + _rotate_half(k) * sin
